+++ preprocessors/Marlin_laser_Spindle_pin.py
@@ -50,7 +50,8 @@ class Marlin_laser_Spindle_pin(PreProc):
         gcode += ';X range: ' + '{: >9s}'.format(xmin) + ' ... ' + '{: >9s}'.format(xmax) + ' ' + units + '\n'
         gcode += ';Y range: ' + '{: >9s}'.format(ymin) + ' ... ' + '{: >9s}'.format(ymax) + ' ' + units + '\n\n'
 
-        gcode += ';Laser Power (Spindle Speed): ' + str(p['spindlespeed']) + '\n' + '\n'
+        gcode += ';Laser Power (Spindle Speed): %s\n' % str(p['spindlespeed'])
+        gcode += ';Laser Minimum Power: %s\n\n' % str(p['laser_min_power'])
 
         gcode += 'G20\n' if p.units.upper() == 'IN' else 'G21\n'
         gcode += 'G90'
@@ -61,24 +62,32 @@ class Marlin_laser_Spindle_pin(PreProc):
         return ''
 
     def lift_code(self, p):
-        gcode = 'M400\n'
-        gcode += 'M5'
-        return gcode
+        if float(p.laser_min_power) > 0.0:
+            return 'M3 S%s' % str(p.laser_min_power)
+        else:
+            if float(p.laser_min_power) > 0.0:
+                return 'M3 S%s' % str(p.laser_min_power)
+            else:
+                gcode = 'M400\n'
+                gcode += 'M5'
+                return gcode
 
     def down_code(self, p):
-        sdir = {'CW': 'M3', 'CCW': 'M4'}[p.spindledir]
         if p.spindlespeed:
-            return '%s S%s' % (sdir, str(p.spindlespeed))
+            return 'M3 S%s' % str(p.spindlespeed)
         else:
-            return sdir
+            return 'M3'
 
     def toolchange_code(self, p):
         return ''
 
     def up_to_zero_code(self, p):
-        gcode = 'M400\n'
-        gcode += 'M5'
-        return gcode
+        if float(p.laser_min_power) > 0.0:
+            return 'M3 S%s' % str(p.laser_min_power)
+        else:
+            gcode = 'M400\n'
+            gcode += 'M5'
+            return gcode
 
     def position_code(self, p):
         # formula for skewing on x for example is:
@@ -133,6 +142,9 @@ class Marlin_laser_Spindle_pin(PreProc):
         return ''
 
     def spindle_stop_code(self, p):
-        gcode = 'M400\n'
-        gcode += 'M5'
-        return gcode
+        if float(p.laser_min_power) > 0.0:
+            return 'M3 S%s' % str(p.laser_min_power)
+        else:
+            gcode = 'M400\n'
+            gcode += 'M5'
+            return gcode